move 3 files from pages to components
