+++ frontend/src/App.jsx
@@ -5,9 +5,10 @@ import './App.css'
 import { ToastContainer, toast } from 'react-toastify';
 
 import { Routes , Route} from 'react-router-dom';
-import CoachRegistration from './pages/CoachRegistration';
-import CoachDetails from './pages/CoachDetails';
-import CoachWaitForApproval from './pages/CoachWaitForApproval';
+import CoachRegistration from './components/CoachRegistration';
+import CoachDetails from './components/CoachDetails';
+import CoachWaitForApproval from './components/CoachWaitForApproval';
+
 
 function App() {
