+++ frontend/src/components/CoachDetails.jsx
